1. Fix unit tests. 2. Do not emit vzeroupper in autovectorized code
We still can not get rid of STACK_FRAME_NON_STANDARD() declarations -
objtool still doesn't understand the direct jumps. But it does
understand jump tables, so use __annotate_jump_table wherever possible.
