+++ tls/t/Makefile
@@ -1,4 +1,4 @@
-# Copyright (C) 2020 Tempesta Technologies, Inc.
+# Copyright (C) 2020-2021 Tempesta Technologies, Inc.
 #
 # This program is free software; you can redistribute it and/or modify it
 # under the terms of the GNU General Public License as published by
@@ -63,7 +63,7 @@ generate_tables: $(AUTOGEN_TARGETS) $(GENERATORS)
 	./tgen_ec256 > ../ecp256_G.autogen.h
 
 # 3). Compile the tests with autogenerated tables.
-$(TARGETS): % : $(ASM-OBJ) %.o
+$(TARGETS): % : $(ASM-OBJ) %.o ttls_mocks.o
 	$(CC) $(CFLAGS) -o $@ $^
 
 $(ASM-OBJ): ../$(subst .o,.S,$(ASM-OBJ))
